+++ internal/model/auth.go
@@ -10,7 +10,7 @@ type User struct {
 	Name      string       `db:"username"`
 	Email     string       `db:"email"`
 	Password  string       `db:"password"`
-	Role      int8         `db:"role"`
+	Role      int          `db:"role"`
 	CreatedAt time.Time    `db:"created_at"`
 	UpdatedAt sql.NullTime `db:"updated_at"`
 }
